% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models.R
\name{compare_models}
\alias{compare_models}
\title{Compare two mixed models using AIC and BIC}
\usage{
compare_models(model1, model2)
}
\arguments{
\item{model1}{A fitted \code{lmerMod} object.}

\item{model2}{A fitted \code{lmerMod} object.}
}
\value{
A data frame with columns:
\describe{
\item{model}{Model label}
\item{AIC}{Akaike Information Criterion}
\item{BIC}{Bayesian Information Criterion}
\item{preferred_AIC}{Logical, TRUE if this model has lower AIC}
\item{preferred_BIC}{Logical, TRUE if this model has lower BIC}
}
}
\description{
Returns a data frame comparing two fitted lme4 models on AIC and BIC,
and identifies which model is preferred by each criterion.
}
\examples{
library(lme4)
data("jedi")
m1 = lmer(darkness ~ anger + (1 | dojo_id), data = jedi)
m2 = lmer(darkness ~ anger + emotional_bonds + (1 | dojo_id), data = jedi)
compare_models(m1, m2)

}
